backend/requirements.txt
# ============================================
# TO THE MOON - Python Dependencies
# ============================================

# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

# Stripe Payments
stripe==7.8.0

# HTTP Client (for external APIs)
requests==2.31.0
httpx==0.26.0

# Validation & Serialization
marshmallow==3.20.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==21.0.0

# Linting & Formatting
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Type Checking
mypy==1.7.1
types-requests==2.31.0.10
